requirements.txt
configparser==3.5.0
olefile==0.44
Pillow==4.1.1
